bad stuff so sad

+++ include/jh.h
@@ -15,17 +15,17 @@ class MCstats
 {
 public:
   // Pointer to a series of pointers which is the location of our fission sites
-  float ** m_fission_sites;
+  double ** m_fission_sites;
   // fission_sites();
-  MCstats(float a_x, float a_y, int grains);
+  MCstats(double a_x, double a_y, int grains);
   // Sets the size of the array of fission sites, points to [0,0]
   void setFissionSite(neutron a_neutron);
   void setFissionSite(neutron a_neutron, int a_n);
-  float ** allocate(int a_row, int a_col);
+  double ** allocate(int a_row, int a_col);
   void deallocate();
-  float * getRow(int row);
+  double * getRow(int row);
   bool fisCon(std::vector<neutron>);
-  void setDims(float a_x, float a_y, int grains);
+  void setDims(double a_x, double a_y, int grains);
   void printFissionSites();
   void normalizeSites();
   int getTotalFissions();
